+++ src/main/java/racingcar/input/utils/carname/validators/exceptions/DuplicateCarNamesException.java
@@ -1,4 +1,4 @@
-package racingcar.input.utils.carname.exception;
+package racingcar.input.utils.carname.validators.exceptions;
 
 public class DuplicateCarNamesException extends RuntimeException {
     public DuplicateCarNamesException(String message) {
