+++ src/main/java/racingcar/input/utils/carname/validators/exceptions/InvalidCarNameFormatException.java
@@ -1,4 +1,4 @@
-package racingcar.input.utils.carname.exception;
+package racingcar.input.utils.carname.validators.exceptions;
 
 public class InvalidCarNameFormatException extends RuntimeException {
     public InvalidCarNameFormatException(String message) {
