+++ src/main/java/racingcar/input/utils/carname/validators/exceptions/LessThanTwoCarNamesException.java
@@ -1,4 +1,4 @@
-package racingcar.input.utils.carname.exception;
+package racingcar.input.utils.carname.validators.exceptions;
 
 public class LessThanTwoCarNamesException extends RuntimeException {
     public LessThanTwoCarNamesException(String message) {
