+++ src/main/java/racingcar/input/utils/racingtrytime/validators/exceptions/NotNaturalNumberRacingTimeException.java
@@ -1,4 +1,4 @@
-package racingcar.input.utils.racingtrytime.exception;
+package racingcar.input.utils.racingtrytime.validators.exceptions;
 
 public class NotNaturalNumberRacingTimeException extends RuntimeException {
     public NotNaturalNumberRacingTimeException(String message) {
