+++ src/test/java/racingcar/input/utils/carname/CarNameFormatValidatorUtilsTest.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.input.utils.carname.exception.InvalidCarNameFormatException;
+import racingcar.input.utils.carname.validators.exceptions.InvalidCarNameFormatException;
 import racingcar.input.utils.carname.validators.CarNameFormatValidatorUtils;
 
 public class CarNameFormatValidatorUtilsTest {
