+++ src/test/java/racingcar/input/utils/carname/CarNamesNotDuplicateValidatorUtilsTest.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.input.utils.carname.exception.DuplicateCarNamesException;
+import racingcar.input.utils.carname.validators.exceptions.DuplicateCarNamesException;
 import racingcar.input.utils.carname.validators.CarNamesNotDuplicateValidatorUtils;
 
 public class CarNamesNotDuplicateValidatorUtilsTest {
