+++ src/test/java/racingcar/input/utils/carname/TwoOrMoreCarNamesValidatorUtilsTest.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.input.utils.carname.exception.LessThanTwoCarNamesException;
+import racingcar.input.utils.carname.validators.exceptions.LessThanTwoCarNamesException;
 import racingcar.input.utils.carname.validators.TwoOrMoreCarNamesValidatorUtils;
 
 public class TwoOrMoreCarNamesValidatorUtilsTest {
